Create core, refactor ABCs
